created User Model Test Case
